fix: re-propose discount calculation files to trigger CI build

+++ TestingApi/Models/DiscountRequest.cs
@@ -5,7 +5,7 @@ namespace TestingApi.Models
     public class DiscountRequest
     {
         [Required(ErrorMessage = "Amount is required.")]
-        [Range(0.01, double.MaxValue, ErrorMessage = "Amount must be greater than zero.")]
+        [Range(0.01, (double)decimal.MaxValue, ErrorMessage = "Amount must be greater than zero.")]
         public decimal Amount { get; set; }
     }
 }
